Move source to src dir
* Add settings file
* Update readme
* Add .env.example
