+++ .env.example
@@ -1,0 +1,8 @@
+BOT_ERTAOZ_TOKEN=
+
+WEATHER_API_ID=
+THECATAPI_API_KEY=
+CORONA_API_ID=
+
+IMGFLIP_API_USERNAME=
+IMGFLIP_API_PASSWORD=
