--- bots/__init__.py
@@ -1,4 +1,0 @@
-from environs import Env
-
-env = Env()
-env.read_env()
