--- system/ertaoz-bot.service
@@ -1,14 +1,0 @@
-[Unit]
-Description=ertaoz.py
-After=network.target
-[Service]
-Type=simple
-User=root
-Group=root
-WorkingDirectory=/srv/telegram_bots
-ExecStart=/srv/bot_env/bin/python main.py --bot ertaoz_bot
-Restart=always
-StandardOutput=syslog
-StandardError=syslog
-[Install]
-WantedBy=multi-user.target
